+++ packages/react-native-messaging-core/ios/Sources/NablaMessagingClient/NablaMessagingClientModule.m
@@ -23,6 +23,10 @@ RCT_EXTERN_METHOD(deleteMessage: (id)messageIdMap
                  conversationId: (id)conversationIdMap
                        callback: (RCTResponseSenderBlock)callback)
 
+RCT_EXTERN_METHOD(retrySendingMessage: (id)messageIdMap
+                       conversationId: (id)conversationIdMap
+                             callback: (RCTResponseSenderBlock)callback)
+
 RCT_EXTERN_METHOD(markConversationAsSeen: (id)conversationIdMap
                                 callback: (RCTResponseSenderBlock)callback)